move webpack plugins and loaders to their own folders
