+++ packages/saber/src/webpack/loaders/toml-loader.js
@@ -1,4 +1,4 @@
-const toml = require('../utils/toml.min')
+const toml = require('../../utils/toml.min')
 
 module.exports = function(source) {
   return JSON.stringify(toml.parse(source))
