+++ packages/saber/src/webpack/loaders/yaml-loader.js
@@ -1,4 +1,4 @@
-const yaml = require('../utils/yaml.min')
+const yaml = require('../../utils/yaml.min')
 
 module.exports = function(source) {
   return JSON.stringify(yaml.safeLoad(source))
